==> Euroxx/5_Diagnostic.ipynb <==
import sys
import os
# Append the library path to PYTHONPATH, so library can be imported.
sys.path.append(os.path.dirname(os.getcwd()))
import itertools
import datetime

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

from library import common as cm

%run setup.py
%matplotlib inline
%load_ext autoreload
%autoreload 2

sns.set(style='darkgrid')

cols = [
    'FREQ=1H_HALFMONEY=otm_MINM=0.8_MAXM=1.5_MINTAU=0_WINDOW=150D_AGGSIDE=False_MATCHING=0.1H_CLOSENESS=False',
    'FREQ=1D_HALFMONEY=otm_MINM=0.8_MAXM=1.5_MINTAU=0_WINDOW=150D_AGGSIDE=False_MATCHING=0.1H_CLOSENESS=False',
    'FREQ=2D_HALFMONEY=otm_MINM=0.8_MAXM=1.5_MINTAU=0_WINDOW=150D_AGGSIDE=False_MATCHING=0.1H_CLOSENESS=False',
]

# Cols are different data sets, rows are different experiments.
rows = [#'Regression/No_Hedge', 
        'Regression/BS_Benchmark', 
        #'Regression/Fixed_Constants',
        #'Regression/Delta_only', 'Regression/Vega_only', 'Regression/Gamma_only', 'Regression/Vanna_only',
        #'Regression/Delta_Gamma', 'Regression/Delta_Vega', 
        #'Regression/Delta_Vanna', 
        #'Regression/Delta_Vega_Gamma', 'Regression/Delta_Vega_Vanna', 'Regression/Delta_Gamma_Vanna',
        #'Regression/Delta_Vega_Gamma_Vanna',
        #'Regression/Hull_White', 'Regression/Hull_White_relaxed',
        #'Network/Normal_Feature', 
        'Network/Normal_Feature_CDF',
        #'Network/Delta_Vega', 
        'Network/Delta_Vega_CDF',
        #'Network/Delta_Vega_Vanna',
        'Network/Delta_Vega_Vanna_CDF',
        #'Network/Spot_Strike_2'
       ]

def run_save_tables(dirs, agg, measure, op_type):
    df_res = inspector.evalPnls(df_dirs, aggregating=aggregating, measure=measure, op_type=op_type)
    if op_type == None:
        file = f'{DATA_DIR}Result/call_and_put_{measure}.csv'
    else:
        file = f'{DATA_DIR}Result/{op_type}_{measure}.csv'
    with open(file, 'w+') as f:
        f.write(f'{datetime.datetime.now()}\n{op_type}_{measure}\n')
    (df_res).to_csv(file, mode='a') 
    return df_res

df_dirs = pd.DataFrame(columns=cols, index=rows)

for x, y in list(itertools.product(rows, cols)):
    df_dirs.loc[x, y] = f'{DATA_DIR}Result/{y}/{x}/'

inspector = cm.Inspector()

aggregating = 'mean'
measure = 'mse'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='call')

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='put')

aggregating = 'mean'
measure = 'median'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)
run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='call')
run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='put')

aggregating = 'mean'
measure = 'lower5%VaR'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='call')

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='put')

aggregating = 'mean'
measure = 'upper95%VaR'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='call')

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='put')

aggregating = 'mean'
measure = 'mean'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)
run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='call')
run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='put')
